+++ services/meta/data.go
@@ -37,6 +37,8 @@ type Data struct {
 
 	MaxShardGroupID uint64
 	MaxShardID      uint64
+
+	DefaultRetentionPolicyName string `json:"-"`
 }
 
 // NewShardOwner sets the owner of the provided shard to the data node
@@ -133,9 +135,7 @@ func (data *Data) RetentionPolicy(database, name string) (*RetentionPolicyInfo, 
 // Returns an error if name is blank or if a database does not exist.
 func (data *Data) CreateRetentionPolicy(database string, rpi *RetentionPolicyInfo) error {
 	// Validate retention policy.
-	if rpi.Name == "" {
-		return ErrRetentionPolicyNameRequired
-	} else if rpi.ReplicaN < 1 {
+	if rpi.ReplicaN < 1 {
 		return ErrReplicationFactorTooLow
 	}
 
@@ -155,9 +155,18 @@ func (data *Data) CreateRetentionPolicy(database string, rpi *RetentionPolicyInf
 		return nil
 	}
 
+	// Determine the retention policy name if it is blank.
+	rpName := rpi.Name
+	if rpName == "" {
+		if data.DefaultRetentionPolicyName == "" {
+			return ErrRetentionPolicyNameRequired
+		}
+		rpName = data.DefaultRetentionPolicyName
+	}
+
 	// Append copy of new policy.
 	rp := RetentionPolicyInfo{
-		Name:               rpi.Name,
+		Name:               rpName,
 		Duration:           rpi.Duration,
 		ReplicaN:           rpi.ReplicaN,
 		ShardGroupDuration: rpi.ShardGroupDuration,
@@ -252,6 +261,13 @@ func (data *Data) UpdateRetentionPolicy(database, name string, rpu *RetentionPol
 
 // SetDefaultRetentionPolicy sets the default retention policy for a database.
 func (data *Data) SetDefaultRetentionPolicy(database, name string) error {
+	if name == "" {
+		if data.DefaultRetentionPolicyName == "" {
+			return ErrRetentionPolicyNameRequired
+		}
+		name = data.DefaultRetentionPolicyName
+	}
+
 	// Find database and verify policy exists.
 	di := data.Database(database)
 	if di == nil {
@@ -728,6 +744,13 @@ type DatabaseInfo struct {
 
 // RetentionPolicy returns a retention policy by name.
 func (di DatabaseInfo) RetentionPolicy(name string) *RetentionPolicyInfo {
+	if name == "" {
+		if di.DefaultRetentionPolicy == "" {
+			return nil
+		}
+		name = di.DefaultRetentionPolicy
+	}
+
 	for i := range di.RetentionPolicies {
 		if di.RetentionPolicies[i].Name == name {
 			return &di.RetentionPolicies[i]